Optimization attempt

+++ MSDF Font Library/FontAtlas/JsonAtlas.cs
@@ -4,7 +4,7 @@
     {
         public string Type { get; set; }
         public int DistanceRange { get; set; }
-        public int Size { get; set; }
+        public float Size { get; set; }
         public int Width { get; set; }
         public int Height { get; set; }
         public string YOrigin { get; set; }
